màj du fichier requestingTranskribus.py après ajout du controle des statuts fournis
